Fix errors on nftmarket.js for old nft own buys
Buying an NFT involves two events, and the parseNftBuy expects that. This is a fix to that.

However this resulted in a problem happening when someone in the past (not the case anymore) would buy an NFT of its own listing. 

In this case, the events would be only 1, and the part of the function that needs the second item of the array would fail.